+++ src/client/components/DemographicsChart.tsx
@@ -19,6 +19,7 @@ const DemographicsChart = ({
 }: {
   readonly demographics: { readonly [id: string]: number };
 }) => {
+  // Only showing hard-coded core metrics here for space reasons, so we can hard-code population as well
   const percentages = mapValues(
     demographics,
     (population: number) =>