+++ CHANGELOG.md
@@ -7,14 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+### Changed
+### Fixed
+
+## [1.11.0] - 2021-10-11
+### Added
+- Add support for demographic groups, to support VAP / CVAP fields [#1034](https://github.com/PublicMapping/districtbuilder/pull/1034)
+- Added bulk-reprocess-regions command [#1022](https://github.com/PublicMapping/districtbuilder/pull/1022)
+- Import using a project template [#1033](https://github.com/PublicMapping/districtbuilder/pull/1033)
 
 ### Changed
-
-### Fixed
+- Don't require AWS credentials to load public S3 assets [#1036](https://github.com/PublicMapping/districtbuilder/pull/1036)
+- Blocks and blockgroups are now populated with FIPS code in name field [#1022](https://github.com/PublicMapping/districtbuilder/pull/1022)
+- The process-geojson command now returns non-zero exit codes on error [#1022](https://github.com/PublicMapping/districtbuilder/pull/1022)
+- The process-geojson, publish-region, and update-region commands now include the input file in the output [#1022](https://github.com/PublicMapping/districtbuilder/pull/1022)
 
 ## [1.10.1] - 2021-09-29
 ### Fixed
 - Fixed project import [#1029](https://github.com/PublicMapping/districtbuilder/pull/1029)
+- Copy project template id when duplicating project [#1037](https://github.com/PublicMapping/districtbuilder/pull/1037)
 
 ## [1.10.0] - 2021-09-28
 ### Added
@@ -340,7 +351,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.10.1...HEAD
+[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.11.0...HEAD
+[1.11.0]: https://github.com/publicmapping/districtbuilder/compare/1.10.1...1.11.0
 [1.10.1]: https://github.com/publicmapping/districtbuilder/compare/1.10.0...1.10.1
 [1.10.0]: https://github.com/publicmapping/districtbuilder/compare/1.9.0...1.10.0
 [1.9.0]: https://github.com/publicmapping/districtbuilder/compare/1.8.0...1.9.0